一些修改

+++ ruoyi-modules/zlyyh-admin/src/main/java/com/ruoyi/zlyyhadmin/service/impl/ProductGroupServiceImpl.java
@@ -144,14 +144,17 @@ public class ProductGroupServiceImpl implements IProductGroupService {
         //TODO 做一些数据校验,如唯一约束
     }
 
+
     /**
      * 批量删除商品组规则配置
      */
     @Override
     public Boolean deleteWithValidByIds(Collection<Long> ids, Boolean isValid) {
-        if(isValid){
-            //TODO 做一些业务上的校验,判断是否需要校验
+        if (ObjectUtil.isEmpty(ids)) {
+            return false;
         }
+        // 删除对应的关联信息
+        productGroupConnectMapper.delete(new LambdaQueryWrapper<ProductGroupConnect>().in(ProductGroupConnect::getProductGroupId, ids));
         return baseMapper.deleteBatchIds(ids) > 0;
     }
 }
